Changed custom notifications to handle multiple instances
