+++ app/src/main/java/org/gem/indo/dooit/api/responses/CustomNotificationResponse.java
@@ -2,49 +2,65 @@ package org.gem.indo.dooit.api.responses;
 
 import com.google.gson.annotations.SerializedName;
 
+import org.gem.indo.dooit.models.Badge;
+import org.gem.indo.dooit.models.CustomNotification;
 import org.joda.time.DateTime;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 /**
  * Created by Chad Garrett on 2017/02/14.
  */
 
 public class CustomNotificationResponse {
-    @SerializedName("persist")
-    private boolean persist;
+    @SerializedName("data")
+    private List<CustomNotification> notifications = new ArrayList<>();
 
-    @SerializedName("message")
-    private String message;
+    @SerializedName("available")
+    private boolean available;
 
-    @SerializedName("publish_date")
-    private DateTime publishDate;
 
-    @SerializedName("expiration_date")
-    private DateTime expireDate;
-
-    @SerializedName("icon")
-    private String icon_url;
-
-    public boolean shouldPersist() {
-        return persist;
-    }
-
-    public boolean isNotificationActive() {
-        DateTime currentDate = DateTime.now();
-
-        // Checks to see if the notification is active
-        return (currentDate.getMillis() >= publishDate.getMillis()) && (currentDate.getMillis() <= expireDate.getMillis());
-    }
-
-    public String getNotificationMessage() {
-        return message;
-    }
-
-    public String getNotificationIcon() {
-        if (!icon_url.equals(null) || icon_url != "") {
-            return icon_url;
+    public List<CustomNotification> getNotifcations() {
+        if (available) {
+            return notifications;
+        } else {
+            return null;
         }
-        return "";
     }
+
+//    @SerializedName("message")
+//    private String message;
+//
+//    @SerializedName("publish_date")
+//    private DateTime publishDate;
+//
+//    @SerializedName("expiration_date")
+//    private DateTime expireDate;
+//
+//    @SerializedName("icon")
+//    private String icon_url;
+//
+//    public List<CustomNotification> getNotifications() {
+//        return badges;
+//    }
+//
+//    public boolean isNotificationActive() {
+//        DateTime currentDate = DateTime.now();
+//
+//        // Checks to see if the notification is active
+//        return (currentDate.getMillis() >= publishDate.getMillis()) && (currentDate.getMillis() <= expireDate.getMillis());
+//    }
+//
+//    public String getNotificationMessage() {
+//        return message;
+//    }
+//
+//    public String getNotificationIcon() {
+//        if (!icon_url.equals(null) || icon_url != "") {
+//            return icon_url;
+//        }
+//        return "";
+//    }
 }
